move test into another test file
